Profile routing scaffolding ready in front-end

+++ client/src/components/profile/modals/DeletePostModalForm.vue
@@ -14,17 +14,19 @@
 
             <v-card-title class="headline pt-5">Eliminar publicación</v-card-title>
 
-            <v-form class="mt-5 pa-0" @submit.prevent="">
+            <v-form class="mt-4 pa-0" @submit.prevent="">
 
                <p class="px-6">Esta seguro que desea eliminar esta publicación ?</p>
 
                <v-card-actions>
 
-                  <v-btn depressed dark @click="submit()" type="submit" class="mb-2 ml-3 text-capitalize" color="blue lighten-1">
+                  <v-btn depressed dark @click="submit()" type="submit" class="mb-2 ml-3 text-capitalize" v-ripple="false"
+                     color="blue lighten-1">
                      <span class="px-2">Eliminar</span>
                   </v-btn>
 
-                  <v-btn depressed light @click="dialog = false" class="mb-2 ml-3 text-capitalize" color="grey lighten-1">
+                  <v-btn depressed light @click="dialog = false" class="mb-2 ml-3 text-capitalize" v-ripple="false"
+                     color="grey lighten-1">
                      <span class="px-2">Cancelar</span>
                   </v-btn>
 
@@ -34,6 +36,17 @@
          </v-card>
 
       </v-dialog>
+
+      <v-snackbar dark v-model="snackbar" :timeout="5000" color="blue lighten-1">
+
+         La publicación ha sido eliminada exitosamente.
+
+         <template v-slot:action="{ attrs }">
+            <v-btn text v-ripple="false" class="white--text text-capitalize" v-bind="attrs" @click="snackbar = false">Ok</v-btn>
+         </template>
+
+      </v-snackbar>
+
    </div>
 
 </template>
@@ -47,7 +60,8 @@
       data(){
          return {
             loading: false,
-            dialog: false
+            dialog: false,
+            snackbar: false
          }
       },
 
@@ -60,18 +74,21 @@
       methods: {
 
          submit(){
-            this.loading = "blue lighten-1";
-            axios.delete("posts/delete/" + this.postId)
-               .then((response) => {
-                  if(response.data){
-                     console.log(response.data);
-                     this.dialog = false;
-                     this.loading = false;
-                  }
-               })
-               .catch((error) => {
-                  console.log(error);
-               });
+            if(!this.loading){
+               this.loading = "blue lighten-1";
+               axios.delete("posts/delete/" + this.postId)
+                  .then((response) => {
+                     if(response.data){
+                        console.log(response.data);
+                        this.dialog = false;
+                        this.loading = false;
+                        this.snackbar = true;
+                     }
+                  })
+                  .catch((error) => {
+                     console.log(error);
+                  });
+            }
          }
       }
    }
